added metadata for sign in , sign up, blog page
